Chef-13:  tweaks to rubygems source option for urls
By default now we don't pass any --source option or set the
`Chef::Config[:rubygems_url]` to anything.  This is so that
if there is external config for rubygems sources we will simply
have rubygems pick it up and use it (the resource acts like
the command line).

There is also no more magic around `--clear-sources` and if you
set the property it sets the flag.

If the `Chef::Config[:rubygems_url]` setting is set then that
becomes the default `--source` flag sent to the command.  This
also supports Arrays now.

If the `source` option is set to a URI (it can also be set to a
filesystem path for poorly-designed-API reasons) then we do not
use the default values at all.  This may be an API break but this
much more clearly allows users to override individual resources
and is more declarative.  You get what you type.

Signed-off-by: Lamont Granquist <lamont@scriptkiddie.org>